# docs/gpt-actions-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SourceNerve GPT Actions",
    "version": "0.1.1",
    "description": "ChatGPT Custom GPT Actions surface for SourceNerve. Git remains the source of truth. Mutations preserve SourceNerve exact-HEAD, file-hash, reviewed-diff, branch and provider guards. The stable MCP extension bridge lets ChatGPT discover and dispatch newly installed downstream MCP tools without changing the GPT Actions schema."
  },
  "servers": [
    {
      "url": "__SOURCENERVE_BASE_URL__",
      "description": "The public SourceNerve origin that served this schema"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/snapshot": {
      "post": {
        "operationId": "getRepositorySnapshot",
        "summary": "Get the current Git HEAD and clean/dirty state",
        "requestBody": {
          "$ref": "#/components/requestBodies/WorkspaceArg"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonResponse"
          }
        }
      }
    },
    "/api/v1/read": {
      "post": {
        "operationId": "readFile",
        "summary": "Read a repository-relative UTF-8 source range and full-file SHA-256",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadFileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonResponse"
          }
        }
      }
    },
    "/api/v1/patch/preview": {
      "post": {
        "operationId": "previewPatch",
        "summary": "Validate a patch without modifying repository files",
        "description": "Requires exact Git HEAD and one expected full-file SHA per touched path. SourceNerve runs the existing guarded patch preview contract.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonResponse"
          }
        }
      }
    },
    "/api/v1/patch/apply": {
      "post": {
        "operationId": "applyPatch",
        "summary": "Apply an already reviewable patch under SourceNerve concurrency guards",
        "description": "Mutates repository files only when exact HEAD, per-file SHA expectations and git apply validation still match.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/JsonResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Use the same server-side SourceNerve bearer token configured by SOURCENERVE_BEARER_TOKEN. Configure it as an API key with Bearer authentication in the GPT Actions editor."
      }
    },
    "responses": {
      "JsonResponse": {
        "description": "SourceNerve JSON response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "requestBodies": {
      "WorkspaceArg": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/WorkspaceArg"
            }
          }
        }
      }
    },
    "schemas": {
...
      "WorkspaceArg": {
        "type": "object",
        "required": [
          "workspace"
        ],
        "properties": {
          "workspace": {
            "type": "string",
            "description": "Configured SourceNerve workspace ID"
          }
        },
        "additionalProperties": false
      },
      "ReadFileRequest": {
        "type": "object",
        "required": [
          "workspace",
          "path"
        ],
        "properties": {
          "workspace": {
            "type": "string"
          },
          "path": {
            "type": "string",
            "description": "Repository-relative path only"
          },
          "start_line": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 1
          },
          "end_line": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 1
          }
        },
        "additionalProperties": false
      },
      "FileExpectation": {
        "type": "object",
        "required": [
          "path",
          "sha256"
        ],
        "properties": {
          "path": {
            "type": "string",
            "description": "Repository-relative touched path"
          },
          "sha256": {
            "type": [
              "string",
              "null"
            ],
            "description": "Full-file SHA-256 returned by readFile; null only when the path is expected not to exist"
          }
        },
        "additionalProperties": false
      },
      "PatchRequest": {
        "type": "object",
        "required": [
          "workspace",
          "expected_head",
          "expected_files",
          "patch"
        ],
        "properties": {
          "workspace": {
            "type": "string"
          },
          "expected_head": {
            "type": "string",
            "description": "Exact Git HEAD captured before generating the patch"
          },
          "expected_files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileExpectation"
            },
            "minItems": 1
          },
          "patch": {
            "type": "string",
            "description": "Unified git patch; bounded by SourceNerve"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}